requirements.txt
asgiref==3.8.1
Authlib==1.3.2
black==24.8.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
dj-database-url==2.2.0
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.9
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.3
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dotenv==0.21.1
requests==2.32.3
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.3
